pyproject.toml
[tool.poetry]
name = "papis-extract"
version = "0.2.0"
description = ""
authors = ["Marty Oehme <marty.oehme@gmail.com>"]
license = "GPL-3.0-or-later"
readme = "README.md"
packages = [{include = "papis_extract"}]
homepage = "https://git.martyoeh.me/Marty/papis-extract"
repository = "https://github.com/marty-oehme/papis-extract"
keywords = ["papis", "pdf", "biblatex", "bibtex", "bibliography", "reference manager", "research", "science"]

[tool.poetry.dependencies]
python = "^3.11"
pymupdf = "^1.23.1"
levenshtein = "^0.21.1"
papis = "^0.13"
click = "^8.1.7"
whoosh = { version = "^2.7.4", optional = true }
python-magic = "^0.4.27"
chevron = "^0.14.0"
beautifulsoup4 = { version = "^4.12.3", optional = true }

[tool.poetry.extras]
whoosh = ["whoosh"]
pocketbook = ["beautifulsoup4"]

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
coverage = "^7.3.1"
pytest-cov = "^4.1.0"

[tool.poetry.plugins."papis.command"]
extract = "papis_extract:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
